Исправил ошибки в ДЗ

+++ src/com/startjava/lesson_2_3_4/calculator/Calculator.java
@@ -3,15 +3,15 @@ package com.startjava.lesson_2_3_4.calculator;
 public class Calculator {
     
     private String[] mathOperation;
-    private double number1;
-    private double number2;
+    private int number1;
+    private int number2;
     private char mathOperatore;
     private double result;
 
     public double calculate(String input) {
         String[] mathOperation = input.split(" ");
-        number1 = Double.parseDouble(mathOperation[0]);
-        number2 = Double.parseDouble(mathOperation[2]);
+        number1 = Integer.parseInt(mathOperation[0]);
+        number2 = Integer.parseInt(mathOperation[2]);
         mathOperatore = mathOperation[1].charAt(0);
 
         result = 1;
